[#59] README 예제를 ReadmeCapture로 단순화
